requirements.txt
2captcha-python==1.1.2
appdirs==1.4.4
APScheduler==3.6.3
argcomplete @ file:///tmp/build/80754af9/argcomplete_1618920853202/work
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
cachetools==4.2.2
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636542104204/work
charset-normalizer==2.0.7
cloudscraper==1.2.58
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
debugpy @ file:///C:/ci/debugpy_1629222823515/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
h11==0.12.0
httplib2==0.20.2
idna==3.3
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916842152/work
ipykernel @ file:///C:/ci/ipykernel_1633527406033/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1635944342510/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1611315820849/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1630690655192/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537209059/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528383308/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///C:/ci/mistune_1594373272338/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479160228/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1635393701858/work
outcome==1.1.0
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyOpenSSL==21.0.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1636093257833/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-telegram-bot==13.11
python-utils==2.5.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==228
pywinpty==0.5.7
pyzmq @ file:///C:/ci/pyzmq_1628276101517/work
requests==2.26.0
requests-toolbelt==0.9.1
selenium==4.0.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///tmp/build/80754af9/traitlets_1632522747050/work
trio==0.19.0
trio-websocket==0.9.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webdriver-manager==3.5.2
webencodings==0.5.1
wincertstore==0.2
wsproto==1.0.0
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
